+++ src/pages/Home/component/Lighter.tsx
